inference/requirements.txt
azure-functions
azure-storage-blob
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.5.0+cpu
torchvision==0.6.0+cpu